+++ Base.Model/Models/Sustancia.cs
@@ -19,7 +19,7 @@ namespace Base.Model.Models
         /*-------------------------------------------
          * PROPIEDADES DE CONSILTA
          * ------------------------------------------*/
-        public long NoCcite { get; set; }
+        public int NoCcite { get; set; }
         public string CodigoSeguridad { get; set; }
